+++ packages/core/src/index.ts
@@ -10,14 +10,7 @@ import * as themes from './themes'
 
 export * from './styled'
 
-export {
-  styled,
-  Theme,
-  RenderChild,
-  themes,
-  reset,
-  fonts,
-}
+export {styled, Theme, RenderChild, themes, reset, fonts}
 
 export * from './primitive'
 export * from './spacer'
@@ -40,3 +33,8 @@ export * from './block'
 export * from './recaptcha'
 export * from './recaptcha-field'
 export * from './link'
+export * from './tumbler'
+export * from './breaker'
+export * from './striper'
+export * from './accordion'
+export * from './grid'