+++ packages/core/src/option/index.ts
@@ -2,10 +2,12 @@ import OptionControl from './OptionControl'
 import OptionModel from './OptionModel'
 import CheckboxControl from './CheckboxControl'
 import RadioControl from './RadioControl'
+import {SwitchControl} from './SwitchControl'
 
 export {
   OptionControl,
   OptionModel,
   CheckboxControl,
   RadioControl,
+  SwitchControl,
 }