+++ packages/desktop/src/notice/Notice.tsx
@@ -1,0 +1,42 @@
+import React, {FC, ReactNode} from 'react'
+import {Flex, FlexItem, Spacer} from '@qiwi/pijma-core'
+import {Heading, Paragraph} from '../typography/'
+
+export interface NoticeProps {
+  icon: ReactNode
+  title?: ReactNode
+}
+
+export const Notice: FC<NoticeProps> = ({icon, title, children}) => (
+  <Flex>
+    <FlexItem
+      shrink={0}
+      basis={6}
+      mr={title ? 4 : 2}
+      children={icon}
+    />
+    {title || children ? (
+      <FlexItem>
+        <Spacer size="xs">
+          {title ? (
+            <Heading
+              size="4"
+              children={title}
+            />
+          ) : (
+            null
+          )}
+          {children ? (
+            <Paragraph
+              children={children}
+            />
+          ) : (
+            null
+          )}
+        </Spacer>
+      </FlexItem>
+    ) : (
+      null
+    )}
+  </Flex>
+)