+++ packages/mobile/src/checkbox-field/CheckboxField.tsx
@@ -6,12 +6,15 @@ import {
   Check,
   OptionField,
   OptionFieldItem,
+  Box,
 } from '@qiwi/pijma-core'
 
 import CheckboxFieldProps from './CheckboxFieldProps'
 import CheckboxFieldOptionModel from './CheckboxFieldOptionModel'
 
-const CheckboxField: FunctionComponent<CheckboxFieldProps<CheckboxFieldOptionModel<any>, any>> = (props) => (
+const CheckboxField: FunctionComponent<
+  CheckboxFieldProps<CheckboxFieldOptionModel<any>, any>
+> = props => (
   <CheckboxControl<CheckboxFieldOptionModel<any>, any>
     tabIndex={props.tabIndex}
     options={props.options}
@@ -30,7 +33,6 @@ const CheckboxField: FunctionComponent<CheckboxFieldProps<CheckboxFieldOptionMod
         onFocus={renderProps.onFocus}
         onBlur={renderProps.onBlur}
         onKeyDown={renderProps.onKeyDown}
-        onMouseLeave={renderProps.onMouseLeave}
         children={renderProps.options.map((option, index) => (
           <OptionControl<any>
             key={index}
@@ -38,14 +40,24 @@ const CheckboxField: FunctionComponent<CheckboxFieldProps<CheckboxFieldOptionMod
             value={option.value}
             onClick={option.onClick}
             onMouseEnter={option.onMouseEnter}
+            onMouseLeave={renderProps.onMouseLeave}
             children={(renderOptionProps) => (
               <OptionFieldItem
                 disabled={option.disabled}
-                icon={<Check disabled={option.disabled} checked={option.checked} focused={option.focused}/>}
+                icon={(
+                  <Box width={6}>
+                    <Check
+                      disabled={option.disabled}
+                      checked={option.checked}
+                      focused={option.focused}
+                    />
+                  </Box>
+                )}
                 label={option.label}
                 description={option.description}
                 onClick={renderOptionProps.onClick}
                 onMouseEnter={renderOptionProps.onMouseEnter}
+                onMouseLeave={renderOptionProps.onMouseLeave}
               />
             )}
           />