+++ packages/core/CHANGELOG.md
@@ -1,3 +1,52 @@
+# [@qiwi/pijma-core-v1.50.0](https://github.com/qiwi/pijma/compare/v1.49.0...v1.50.0) (2019-06-14)
+
+
+### Features
+
+* block-content ([74af5fd](https://github.com/qiwi/pijma/commit/74af5fd))
+
+# [@qiwi/pijma-core-v1.48.0](https://github.com/qiwi/pijma/compare/v1.47.0...v1.48.0) (2019-06-10)
+
+
+### Features
+
+* block content ([622fda5](https://github.com/qiwi/pijma/commit/622fda5))
+
+# [@qiwi/pijma-core-v1.47.0](https://github.com/qiwi/pijma/compare/v1.46.0...v1.47.0) (2019-06-10)
+
+
+### Features
+
+* grid ([45e60f4](https://github.com/qiwi/pijma/commit/45e60f4))
+
+# [@qiwi/pijma-core-v1.46.0](https://github.com/qiwi/pijma/compare/v1.45.0...v1.46.0) (2019-06-06)
+
+
+### Features
+
+* add accordion ([bc30234](https://github.com/qiwi/pijma/commit/bc30234))
+
+# [@qiwi/pijma-core-v1.42.0](https://github.com/qiwi/pijma/compare/v1.41.0...v1.42.0) (2019-05-29)
+
+
+### Features
+
+* add switch ([e956066](https://github.com/qiwi/pijma/commit/e956066))
+
+# [@qiwi/pijma-core-v1.41.0](https://github.com/qiwi/pijma/compare/v1.40.0...v1.41.0) (2019-05-29)
+
+
+### Features
+
+* striper ([caf2af8](https://github.com/qiwi/pijma/commit/caf2af8))
+
+# [@qiwi/pijma-core-v1.40.0](https://github.com/qiwi/pijma/compare/v1.39.0...v1.40.0) (2019-05-21)
+
+
+### Features
+
+* breaker ([10bc3c0](https://github.com/qiwi/pijma/commit/10bc3c0))
+
 # [@qiwi/pijma-core-v1.39.0](https://github.com/qiwi/pijma/compare/v1.38.0...v1.39.0) (2019-05-13)